# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OMEDIA",
    "description": "Disk-first media organization API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/settings/providers": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Provider Settings",
        "operationId": "provider_settings",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderStatus"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Save Provider Settings",
        "operationId": "save_provider_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderStatus"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/organize": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Organize Settings",
        "operationId": "organize_settings",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizeSettings"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Save Organize Settings",
        "operationId": "save_organize_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizeSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizeSettings"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/watch-runtime": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Watch Runtime Settings",
        "operationId": "watch_runtime_settings",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchRuntimeSettings"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Save Watch Runtime Settings",
        "operationId": "save_watch_runtime_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchRuntimeSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchRuntimeSettings"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorDetail"
      },
      "ApiErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse"
      },
...
      "MediaExtensionSettings": {
        "properties": {
          "video": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Video"
          },
          "subtitle": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Subtitle"
          },
          "sidecar": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sidecar"
          }
        },
        "type": "object",
        "required": [
          "video",
          "subtitle",
          "sidecar"
        ],
        "title": "MediaExtensionSettings"
      },
...
      "OrganizeSettings": {
        "properties": {
          "extensions": {
            "$ref": "#/components/schemas/MediaExtensionSettings"
          },
          "min_non_subtitle_file_size_mb": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Non Subtitle File Size Mb",
            "default": 50
          }
        },
        "type": "object",
        "required": [
          "extensions"
        ],
        "title": "OrganizeSettings"
      },
...
      "ProviderSettingsUpdate": {
        "properties": {
          "tmdb_api_key": {
            "type": "string",
            "title": "Tmdb Api Key"
          },
          "tmdb_base_url": {
            "type": "string",
            "title": "Tmdb Base Url"
          },
          "tmdb_rate_limit": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Tmdb Rate Limit"
          },
          "tmdb_proxy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tmdb Proxy"
          },
          "llm_api_key": {
            "type": "string",
            "title": "Llm Api Key"
          },
          "llm_base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Base Url"
          },
          "llm_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Model"
          },
          "llm_batch_size": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Llm Batch Size"
          },
          "llm_rate_limit": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Llm Rate Limit"
          },
          "llm_proxy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Proxy"
          }
        },
        "type": "object",
        "required": [
          "tmdb_api_key",
          "tmdb_base_url",
          "tmdb_rate_limit",
          "llm_api_key",
          "llm_batch_size",
          "llm_rate_limit"
        ],
        "title": "ProviderSettingsUpdate"
      },
      "ProviderStatus": {
        "properties": {
          "tmdb_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tmdb Api Key"
          },
          "tmdb_base_url": {
            "type": "string",
            "title": "Tmdb Base Url"
          },
          "tmdb_rate_limit": {
            "type": "number",
            "title": "Tmdb Rate Limit"
          },
          "tmdb_proxy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tmdb Proxy"
          },
          "llm_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Api Key"
          },
          "llm_base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Base Url"
          },
          "llm_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Model"
          },
          "llm_batch_size": {
            "type": "integer",
            "title": "Llm Batch Size"
          },
          "llm_rate_limit": {
            "type": "number",
            "title": "Llm Rate Limit"
          },
          "llm_proxy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Proxy"
          }
        },
        "type": "object",
        "required": [
          "tmdb_base_url",
          "tmdb_rate_limit",
          "llm_batch_size",
          "llm_rate_limit"
        ],
        "title": "ProviderStatus"
      },
...
      "WatchRuntimeSettings": {
        "properties": {
          "poll_interval_seconds": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Poll Interval Seconds"
          },
          "stability_debounce_seconds": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Stability Debounce Seconds"
          }
        },
        "type": "object",
        "required": [
          "poll_interval_seconds",
          "stability_debounce_seconds"
        ],
        "title": "WatchRuntimeSettings"
      },
...
    }
  }
}